Fix String::compareTo(const char*) for invalid strings
When comparing an invalid String object to a non-empty char*, this would
erronously return 0 (equal) because of a typo.

This bug also masked three incorrect checks in related testcases. In two
cases, a String was made invalid and then checked to still contain a
value (these were changed to check that the string is invalid) and in
one case the wrong string was checked.